chore: adaptor vite vue plugin middleware
